Initial commit
This is the initial commit of an old backup from Bitbucket, from back
when it supported Mercurial hosting. I have not imported the history
since it didn't really contain more than a handful of commits at most.
Note that this is an old project from 2015, and hence it does not
consider rescent additions to the C++ language.